Crate grapes

Reactivity for gtk4-rs

Provides reactivity primitives and macros for component approach

Re-exports

pub use css::Css;

pub use reactive::Reactive;

pub use gtk;

pub use gtk::cairo;

pub use gtk::gio;

pub use gtk::glib;

pub use gtk::pango;

pub use layer_shell;

pub use tokio;

pub use component::*;

pub use effect::*;

pub use state::*;

Modules

component

css

effect

extensions

prelude

reactive

state

Structs

Task

Simple RAII wraper for tokio task

Statics

RT

Functions

derived

Creates state which depends on another state

effect

Creates effect

state

Creates reactive state wrapped by Rc

Derive Macros

Component

Generates code that will have to be written anyway

WindowComponent

Generates code that will have to be written anyway