grapes

Module prelude

Re-exports

pub use crate::reactive::Reactive;

pub use crate::Component;

pub use crate::UpdateableComponent;

pub use crate::RT;

pub use gtk;

pub use gtk::gdk;

pub use gtk::glib;

pub use crate::extensions::*;

Enums

UnixSocketAddressPath

Traits

AccessibleExt

AccessibleExtManual

Trait containing manually implemented methods of Accessible.

AccessibleRangeExt

AccessibleTextExt

ActionExt

ActionGroupExt

ActionMapExt

ActionMapExtManual

ActionableExt

ActionableExtManual

Trait containing manually implemented methods of Actionable.

AdjustmentExt

AppChooserExtDeprecated

AppInfoExt

AppInfoExtManual

AppLaunchContextExt

ApplicationCommandLineExt

ApplicationExt

ApplicationExtManual

ApplicationWindowExt

AsyncInitableExt

AsyncResultExt

BoxExt

BufferedInputStreamExt

BufferedOutputStreamExt

BuildableExt

BuilderScopeExt

ButtonExt

CancellableExt

CancellableExtManual

Cast

Upcasting and downcasting support.

CastNone

Convenience trait mirroring Cast, implemented on Option<Object> types.

CellAreaContextExtDeprecated

CellAreaExtDeprecated

CellAreaExtManualDeprecated

Trait containing manually implemented methods of CellArea.

CellEditableExtDeprecated

CellLayoutExtDeprecated

CellLayoutExtManualDeprecated

Trait containing manually implemented methods of CellLayout.

CellRendererExtDeprecated

CellRendererTextExtDeprecated

CheckButtonExt

ColorChooserExtDeprecated

ColorChooserExtManualDeprecated

Trait containing manually implemented methods of ColorChooser.

ComboBoxExtDeprecated

ComboBoxExtManualDeprecated

Trait containing manually implemented methods of ComboBox.

ConstraintTargetExt

ContentProviderExt

ContentProviderExtManual

Trait containing manually implemented methods of ContentProvider.

ConverterExt

ConverterExtManual

ConverterInputStreamExt

ConverterOutputStreamExt

DBusInterfaceExt

DBusInterfaceSkeletonExt

DBusMethodCall

DBusObjectExt

DBusObjectManagerClientExt

DBusObjectManagerClientExtManual

DBusObjectManagerExt

DBusObjectManagerServerExt

DBusObjectProxyExt

DBusObjectSkeletonExt

DBusProxyExt

DBusProxyExtManual

DataInputStreamExt

DataInputStreamExtManual

DataOutputStreamExt

DatagramBasedExt

DatagramBasedExtManual

DesktopAppInfoExtManual

DeviceExt

DevicePadExt

DialogExtDeprecated

DialogExtManualDeprecated

Trait containing manually implemented methods of Dialog.

DisplayExt

DisplayExtManual

Trait containing manually implemented methods of Display.

DragExt

DragSurfaceExt

DragSurfaceExtManual

Trait containing manually implemented methods of DragSurface.

DrawContextExt

DrawContextExtManual

Trait containing manually implemented methods of DrawContext.

DrawingAreaExt

DrawingAreaExtManual

Trait containing manually implemented methods of DrawingArea.

DriveExt

DtlsClientConnectionExt

DtlsConnectionExt

DtlsServerConnectionExt

DynamicObjectRegisterExt

EditableExt

EditableExtManual

Trait containing manually implemented methods of Editable.

EmblemedIconExt

EntryBufferExt

EntryBufferExtManual

Trait containing manually implemented methods of EntryBuffer.

EntryExt

EntryExtManual

Trait containing manually implemented methods of Entry.

ErrorDomain

GLib error domain.

EventControllerExt

EventControllerExtManual

EventKind

A common trait implemented by the various Event types.

FileChooserExtDeprecated

FileChooserExtManualDeprecated

Trait containing manually implemented methods of FileChooser.

FileDescriptorBasedExtManual

FileEnumeratorExt

FileEnumeratorExtManual

FileExt

FileExtManual

FileIOStreamExt

FileInputStreamExt

FileMonitorExt

FileOutputStreamExt

FilterExt

FilterInputStreamExt

FilterOutputStreamExt

FixedExt

FixedSizeVariantType

Trait for fixed size variant types.

FlowBoxChildExt

FontChooserExtDeprecated

FontChooserExtManualDeprecated

Trait containing manually implemented methods of FontChooser.

FontExt

FontFaceExt

FontFamilyExt

FontMapExt

FontsetExt

FrameExt

FromVariant

Extracts a value.

GLAreaExt

GLContextExt

GObjectPropertyExpressionExt

Trait containing convenience methods in creating PropertyExpression that looks up a property of a [glib::Object].

GdkAppLaunchContextExt

GdkCairoContextExt

Trait containing integration methods with [cairo::Context].

GdkCairoSurfaceExt

Trait containing integration methods with [cairo::Surface].

GestureDragExt

GestureExt

GestureSingleExt

GridExt

GskRendererExt

GtkApplicationExt

GtkMountOperationExt

GtkWindowExt

HasParamSpec

IMContextExt

IMMulticontextExt

IOStreamExt

IOStreamExtManual

IconExt

InetAddressExt

InetAddressExtManual

InetAddressMaskExt

InetSocketAddressExt

InitableExt

InputStreamExt

InputStreamExtManual

IsA

Declares the “is a” relationship.

IsAttribute

IsExpression

A common trait implemented by the various Expression types.

IsRenderNode

A common trait implemented by the various RenderNode types.

LayoutChildExt

LayoutManagerExt

ListBoxRowExt

ListItemExt

ListModelExt

ListModelExtManual

LoadableIconExt

MediaFileExt

MediaStreamExt

MediaStreamExtManual

MemoryInputStreamExt

MemoryMonitorExt

MemoryOutputStreamExt

MenuAttributeIterExt

MenuLinkIterExt

MenuModelExt

MonitorExt

MountExt

MountOperationExt

MultiFilterExt

NativeDialogExt

NativeDialogExtManualDeprecated

Trait containing manually implemented methods of NativeDialog.

NativeExt

NetworkAddressExt

NetworkMonitorExt

NetworkServiceExt

ObjectClassExt

Trait for class methods on Object and subclasses of it.

ObjectExt

ObjectType

Implemented by types representing glib::Object and subclasses of it.

OrientableExt

OutputStreamExt

OutputStreamExtManual

PaintableExt

ParamSpecBuilderExt

A trait implemented by the various [ParamSpec] builder types.

ParamSpecType

PermissionExt

PixbufAnimationExt

PixbufAnimationExtManual

PixbufLoaderExt

PollableInputStreamExt

PollableInputStreamExtManual

PollableOutputStreamExt

PollableOutputStreamExtManual

PopoverExt

PopupExt

PrintOperationExt

PrintOperationPreviewExt

ProxyAddressEnumeratorExt

ProxyAddressExt

ProxyExt

ProxyResolverExt

RangeExt

RecentManagerExt

RemoteActionGroupExt

RendererExt

ResolverExt

RootExt

ScaleButtonExt

ScaleExt

ScaleExtManual

Trait containing manually implemented methods of Scale.

ScrollableExt

SeatExt

SectionModelExt

SeekableExt

SelectionModelExt

SettingsBackendExt

SettingsExt

SettingsExtManual

ShortcutActionExt

ShortcutManagerExt

ShortcutTriggerExt

ShortcutTriggerExtManual

Trait containing manually implemented methods of ShortcutTrigger.

SimpleProxyResolverExt

SimpleProxyResolverExtManual

SnapshotExt

SnapshotExtManual

SocketAddressEnumeratorExt

SocketAddressExt

SocketClientExt

SocketConnectableExt

SocketConnectionExt

SocketControlMessageExt

SocketControlMessageExtManual

SocketExt

SocketExtManual

SocketListenerExt

SocketListenerExtManual

SocketServiceExt

SorterExt

StaticType

Types that are supported by GLib dynamic typing.

StaticTypeExt

StaticVariantType

Returns VariantType of Self.

StyleContextExtDeprecated

StyleProviderExt

SurfaceExt

SurfaceExtManual

Trait containing manually implemented methods of Surface.

SymbolicPaintableExt

TcpConnectionExt

TcpWrapperConnectionExt

TextBufferExt

TextBufferExtManual

Trait containing manually implemented methods of TextBuffer.

TextChildAnchorExt

TextMarkExt

TextTagExt

TextViewExt

TextureExt

TextureExtManual

Trait containing manually implemented methods of Texture.

ThreadedSocketServiceExt

TlsBackendExt

TlsCertificateExt

TlsClientConnectionExt

TlsConnectionExt

TlsConnectionExtManual

TlsDatabaseExt

TlsFileDatabaseExt

TlsInteractionExt

TlsPasswordExt

TlsServerConnectionExt

ToSendValue

Converts to SendValue.

ToValue

Trait to convert a value to a Value.

ToVariant

Converts to Variant.

ToggleButtonExt

ToplevelExt

ToplevelExtManual

Trait containing manually implemented methods of Toplevel.

TreeDragDestExtDeprecated

TreeDragSourceExtDeprecated

TreeModelExtDeprecated

TreeModelExtManualDeprecated

Trait containing manually implemented methods of TreeModel.

TreeModelFilterExtDeprecated

TreeModelFilterExtManualDeprecated

Trait containing manually implemented methods of TreeModelFilter.

TreeModelSortExtDeprecated

TreeSortableExtDeprecated

TreeSortableExtManualDeprecated

Trait containing manually implemented methods of TreeSortable.

TreeViewExtDeprecated

TreeViewExtManualDeprecated

Trait containing manually implemented methods of TreeView.

TypeModuleExt

TypePluginExt

UnixConnectionExt

UnixCredentialsMessageExt

UnixFDListExt

UnixFDListExtManual

UnixFDMessageExt

UnixFDMessageExtManual

UnixInputStreamExt

UnixInputStreamExtManual

UnixOutputStreamExt

UnixOutputStreamExtManual

UnixSocketAddressExt

UnixSocketAddressExtManual

ValueType

A type that can be stored in Values.

VfsExt

VolumeExt

VolumeMonitorExt

WidgetExt

WidgetExtManual

Trait containing manually implemented methods of Widget.

WindowGroupExt

Derive Macros

Component

Generates code that will have to be written anyway