grapes::prelude

Trait ToValue

pub trait ToValue {
    // Required methods
    fn to_value(&self) -> Value;
    fn value_type(&self) -> Type;
}

Trait to convert a value to a Value.

Similar to other common conversion traits, the following invariants are guaranteed:

Invertibility: x.to_value().get().unwrap() == x. In words, [FromValue] is the inverse of ToValue.

Idempotence: x.to_value() == x.to_value().to_value(). In words, applying ToValue multiple times yields the same result as applying it once. Idempotence also applies the other way around: value.get::<Value>() is a no-op.

There is also the possibility to wrap values within values, see [BoxedValue]. All (un-)boxing needs to be done manually, and will be preserved under the conversion methods.

The conversion methods may cause values to be cloned, which may result in reference counter changes or heap allocations depending on the source and target type.

Required Methods

fn to_value(&self) -> Value

Convert a value to a Value.

fn value_type(&self) -> Type

Returns the type identifier of self.

This is the type of the value to be returned by to_value.